Everything but the lyrics cog works I think
